add web API support for flashing
