+++ src/js/msp/MSPConnector.js
@@ -6,85 +6,105 @@ import MSP from "../msp";
 import FC from "../fc";
 import MSPCodes from "./MSPCodes";
 import CONFIGURATOR from "../data_storage";
-import serial from "../serial";
+import serial from "../webSerial";
 import { gui_log } from "../gui_log";
 
 /**
  * This seems to be mainly used in firmware flasher parts.
  */
-const MSPConnectorImpl = function () {
-    this.baud = undefined;
-    this.port = undefined;
-    this.onConnectCallback = undefined;
-    this.onTimeoutCallback = undefined;
-    this.onDisconnectCallback = undefined;
-};
+class MSPConnectorImpl {
+    constructor() {
+        this.baud = undefined;
+        this.port = undefined;
+        this.onConnectCallback = undefined;
+        this.onTimeoutCallback = undefined;
+        this.onDisconnectCallback = undefined;
+    }
 
-MSPConnectorImpl.prototype.connect = function (port, baud, onConnectCallback, onTimeoutCallback, onFailureCallback) {
+    connect(port, baud, onConnectCallback, onTimeoutCallback, onFailureCallback) {
 
-    const self = this;
-    self.port = port;
-    self.baud = baud;
-    self.onConnectCallback = onConnectCallback;
-    self.onTimeoutCallback = onTimeoutCallback;
-    self.onFailureCallback = onFailureCallback;
+        this.port = port;
+        this.baud = baud;
+        this.onConnectCallback = onConnectCallback;
+        this.onTimeoutCallback = onTimeoutCallback;
+        this.onFailureCallback = onFailureCallback;
 
-    serial.connect(self.port, {bitrate: self.baud}, function (openInfo) {
-        if (openInfo) {
-            const disconnectAndCleanup = function() {
-                serial.disconnect(function(result) {
-                    console.log('Disconnected');
+        this.connectHandler = (openInfo) => {
+            if (openInfo) {
+                const disconnectAndCleanup = () => {
+                    serial.disconnect((result) => {
+                        console.log('Disconnected', result);
 
-                    MSP.clearListeners();
+                        MSP.clearListeners();
 
-                    self.onTimeoutCallback();
-                });
+                        this.onTimeoutCallback();
+                    });
 
-                MSP.disconnect_cleanup();
-            };
+                    MSP.disconnect_cleanup();
+                };
 
-            FC.resetState();
+                FC.resetState();
 
-            // disconnect after 10 seconds with error if we don't get IDENT data
-            GUI.timeout_add('msp_connector', function () {
-                if (!CONFIGURATOR.connectionValid) {
-                    gui_log(i18n.getMessage('noConfigurationReceived'));
+                // disconnect after 10 seconds with error if we don't get IDENT data
+                GUI.timeout_add('msp_connector', function () {
+                    if (!CONFIGURATOR.connectionValid) {
+                        gui_log(i18n.getMessage('noConfigurationReceived'));
 
-                    disconnectAndCleanup();
+                        disconnectAndCleanup();
+                    }
+                }, 10000);
+
+                function read_serial_adapter(e) {
+                    read_serial(e.detail.buffer);
                 }
-            }, 10000);
 
-            serial.onReceive.addListener(read_serial);
+                serial.removeEventListener('receive', read_serial_adapter);
+                serial.addEventListener('receive', read_serial_adapter);
 
-            const mspHelper = new MspHelper();
-            MSP.listen(mspHelper.process_data.bind(mspHelper));
+                const mspHelper = new MspHelper();
+                MSP.listen(mspHelper.process_data.bind(mspHelper));
 
-            MSP.send_message(MSPCodes.MSP_API_VERSION, false, false, function () {
-                CONFIGURATOR.connectionValid = true;
+                MSP.send_message(MSPCodes.MSP_API_VERSION, false, false, () => {
+                    CONFIGURATOR.connectionValid = true;
 
-                GUI.timeout_remove('msp_connector');
-                console.log('Connected');
+                    GUI.timeout_remove('msp_connector');
+                    console.log('Connected');
 
-                self.onConnectCallback();
-            });
-        } else {
-            gui_log(i18n.getMessage('serialPortOpenFail'));
-            self.onFailureCallback();
-        }
-    });
-};
+                    this.onConnectCallback();
+                });
+            } else {
+                gui_log(i18n.getMessage('serialPortOpenFail'));
+                this.onFailureCallback();
+            }
+        };
 
-MSPConnectorImpl.prototype.disconnect = function(onDisconnectCallback) {
-    self.onDisconnectCallback = onDisconnectCallback;
+        this.disconnectHandler = (detail) => {
+            console.log('Disconnected', detail);
+            MSP.clearListeners();
+            MSP.disconnect_cleanup();
+        };
 
-    serial.disconnect(function (result) {
-        MSP.clearListeners();
-        console.log('Disconnected');
+        serial.removeEventListener('connect', (e) => this.connectHandler(e.detail));
+        serial.addEventListener('connect', (e) => this.connectHandler(e.detail));
 
-        self.onDisconnectCallback(result);
-    });
+        serial.removeEventListener('disconnect', (e) => this.disconnectHandler(e));
+        serial.addEventListener('disconnect', (e) => this.disconnectHandler(e));
 
-    MSP.disconnect_cleanup();
-};
+        serial.connect(this.port, { baudRate: this.baud });
+    }
+
+    disconnect(onDisconnectCallback) {
+        this.onDisconnectCallback = onDisconnectCallback;
+
+        serial.disconnect((result) => {
+            MSP.clearListeners();
+            console.log('Disconnected', result);
+
+            this.onDisconnectCallback(result);
+        });
+
+        MSP.disconnect_cleanup();
+    }
+}
 
 export default MSPConnectorImpl;
